[ARV-66] refactor : repository interface -> domain package안으로 변경
