+++ src/main/java/com/backend/allreva/survey/command/domain/SurveyBoardingDateCommandRepository.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.survey.command.application;
+package com.backend.allreva.survey.command.domain;
 
 import com.backend.allreva.survey.command.domain.Survey;
 import com.backend.allreva.survey.command.domain.SurveyBoardingDate;
