+++ src/test/java/com/backend/allreva/survey/command/application/SurveyCommandIntegrationTest.java
@@ -10,9 +10,7 @@ import com.backend.allreva.survey.command.application.dto.JoinSurveyRequest;
 import com.backend.allreva.survey.command.application.dto.OpenSurveyRequest;
 import com.backend.allreva.survey.command.application.dto.SurveyIdRequest;
 import com.backend.allreva.survey.command.application.dto.UpdateSurveyRequest;
-import com.backend.allreva.survey.command.domain.Survey;
-import com.backend.allreva.survey.command.domain.SurveyBoardingDate;
-import com.backend.allreva.survey.command.domain.SurveyJoin;
+import com.backend.allreva.survey.command.domain.*;
 import com.backend.allreva.survey.command.domain.value.BoardingType;
 import com.backend.allreva.survey.command.domain.value.Region;
 import com.backend.allreva.survey.exception.SurveyInvalidBoardingDateException;
