+++ src/test/java/com/backend/allreva/survey/query/application/SurveyQueryIntegrationTest.java
@@ -5,8 +5,8 @@ import com.backend.allreva.concert.command.domain.ConcertRepository;
 import com.backend.allreva.member.command.domain.Member;
 import com.backend.allreva.member.command.domain.MemberRepository;
 import com.backend.allreva.support.IntegrationTestSupport;
-import com.backend.allreva.survey.command.application.SurveyBoardingDateCommandRepository;
-import com.backend.allreva.survey.command.application.SurveyCommandRepository;
+import com.backend.allreva.survey.command.domain.SurveyBoardingDateCommandRepository;
+import com.backend.allreva.survey.command.domain.SurveyCommandRepository;
 import com.backend.allreva.survey.command.application.SurveyCommandService;
 import com.backend.allreva.survey.command.application.dto.OpenSurveyRequest;
 import com.backend.allreva.survey.command.domain.value.Region;
